HW: added LEDs and AVCC cap
There were no LEDs.
Why were there no LEDs?!
Everything needs LEDs!

So now there are, indicating that the device is actually powered, which
will lessen occasional confusion about that (especially when one forgets
to adjust the jumper after flashing).

But simple LED alone is boring, so there's now two below the IC itself,
giving a great glow to the device.

Note that this is really only just a power indication and not routed to
any output pin, so the LEDs are always on and can't be controlled by the
microcontroller itself -- at least at this point.

Also added 100nF cap to decouple VCC and AVCC, just in case.