FW: moved examples and refactored V-USB config override handling
V-USB is great, but even though this is specifically a Random *USB*
Device, it doesn't have to be, and using V-USB functionality is really
more optional, and is therefore now separated.

Not that there's much to separate it from, but now it's easier to add
bare metal examples in the future.

While moving the examples, the config handling for V-USB was refactored
along the way. It didn't make sense anymore to explicitly undef the
values to override them. Instead, the main usbconfig.h uses values set
in an usbconfig-rudy.h override header file that can be adjusted for
each example project individually.