FW: Added script to check if all required tools are installed
To make life easier, check-system.sh is going to check if the required
tools (make, GCC, libc, binutils, AVRDUDE) are installed on the system
and tells if some are missing. On .deb or Red Hat systems, it also
offers to install missing packages right away.

If all packages are installed, it compiles a little dummy C file to
verify that it really all works.