FW: use correct binutils package for Debian in check-system script
Like avr-gcc / gcc-avr, it's the same with avr-binutils / binutils-avr,
which wasn't handled properly. Now it is, and since there's plenty of
checking /etc/*release now, the script just determines the distro right
away in the beginning.

Also, this must be actual bash script, not sh, as it uses bash-specific
stuff like += operator (fails at Raspberry Pi for example)